backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
aiohttp>=3.11.0
stripe>=11.5.0
python-multipart>=0.0.18
httpx>=0.28.0
PyJWT>=2.10.0
passlib[bcrypt]>=1.7.4
prometheus-fastapi-instrumentator>=7.0.0
pypdf>=5.1.0
python-docx>=1.1.2
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
redis>=5.2.0
langgraph>=0.2.60
langchain-core>=0.3.28
qdrant-client>=1.12.0
chromadb>=0.5.23
crewai>=0.85.0
